backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
langchain
langchain-google-genai
langchain-community
pydantic
